Fix cannot cast object 'false' to Integer
The predefined value of MaxRetries ist 0. Groovy interpret this as a
boolean false instead of an integer.

A call with 0 raised the following exception:

Cannot cast object 'false' with class 'java.lang.Boolean' to class
'java.lang.Integer'

The applied fix, catches a possible boolean value and transforms it to
an integer.